+++ src/containers/stories/south-africa-mozambique/section-1/map/layer-manager/item.tsx
@@ -6,9 +6,10 @@ import { Layer } from "@deck.gl/core";
 
 import { parseConfig } from "@/lib/json-converter";
 import { LAYERS } from "@/lib/layers";
-import { useLocationId } from "@/lib/locations";
 
-import { useSyncDepth, useSyncLocation } from "@/app/(atlas)/atlas/store";
+import { useSyncDepth } from "@/app/(atlas)/atlas/store";
+
+import { useBbox } from "@/containers/stories/south-africa-mozambique/section-1/map/utils";
 
 import DeckLayer from "@/components/map/layers/deck-layer";
 
@@ -20,10 +21,9 @@ interface LayerManagerItemProps {
 
 const LayerManagerItem = ({ id, settings }: LayerManagerItemProps) => {
   const LAYER = LAYERS.find((l) => l.id === id);
-  const [location] = useSyncLocation();
   const [depth] = useSyncDepth();
 
-  const LOCATION = useLocationId(location);
+  const BBOX = useBbox();
 
   const { config, params_config } = LAYER ?? {};
   const c = parseConfig<Layer>({
@@ -36,7 +36,7 @@ const LayerManagerItem = ({ id, settings }: LayerManagerItemProps) => {
     params_config,
     settings: {
       ...settings,
-      extent: LOCATION?.bounds || null,
+      extent: BBOX,
       depth0: depth[0],
       depth1: depth[1],
     },
